runtime: export `_swift_willThrow`
`_swift_willThrow` is meant as a debugging hook and is a symbol that is meant to be exported.  This will not be exported currently on Windows as it is not attributed, and the default visibility is hidden.  Adjust this to make it available on Windows.

Thanks to @grynspan for identifying this issue!