Relax watchos test requirements
To help this test run in production environments, I've reduced the
requirements of the test. It used to require that the compiler was
running on an armv7k or arm64_32 watchOS device, which I don't think is
a test configuration that gets hit all that often.

The new requirements are that you're on watchOS or the simulator, and
can produce arm64_32 code. The test only verifies that the emitted asm
is valid, not that it runs, so we don't need to be running on the
specific hardware. Additionally, the test only checks the output of
arm64_32, not of armvk7, so we only require the 64-bit arm backend.

It would be better if we could check for available SDKs though. The test
only requires an SDK, it doesn't need to be run on the simulator or the
watch.