Document PrunedLiveness to explain guaranteed phis liveness.
In a sense, liveness is wrong for dead guaranteed phis that aren't
enclosed in a separate borrow scope:

     left:           // live-within
       br merge(%v)
     right:          // live-within
       br merge(%v)
     merge(%deadPhi) // dead

But this is the way that liveness layering works. PrunedLiveness
simply doesn't know which phis will "end" liveness. This will be the
job of OwnershipLiveness, which will have safe APIs for situations
where this weirdness matters.