[stdlib] Rework grapheme breaking state machine
Grapheme breaking currently works by essentially looking up the
grapheme break properties of each scalar twice — once when it is on
the right side of a potential break position, and once more when
it is on the left side of the next. This is wasteful.

- Incorporate the last scalar value as well as its grapheme break
   property into state of the grapheme breaking state machine.
- Remove the old grapheme breaking speedup implemented by
   `_hasGraphemeBreakBetween` — having it as a separate phase used
   to be useful when we delegated to ICU, but now that we’ve a
   Swift native implementation, this is no longer worth it.
   However, still prevent dispatching to
   `_swift_stdlib_getGraphemeBreakProperty` for these characters
   by checking for them as part of the property lookup code in Swift.
- Add a quick check for a guaranteed break / no break case between
   two specified scalars, with no information about any preceding
   scalars.

This should measurably speed up `_CharacterRecognizer`, but I expect
it will do the opposite for `String`, as it cannot currently preserve
grapheme breaking state across Character boundaries. (Indices will
need to be reworked to allow this.)

# Conflicts:
#	stdlib/public/core/StringGraphemeBreaking.swift