Disable ASTGen runtime-compat libs
The runtime compat libraries are not available during the stage-0
compiler ASTGen build. As a result, the compiler fails to link because
the compat symbols are not available.

Once we have better stage tracking, we should turn this on for stages
after the compat56 library has been built. Turning it off for now is
safe enough because the ASTGen library doesn't use concurrency, and thus
far, that is the only area with backdeployed fixes.