build: tweak flags further
Because we know that we are building the standard library which requires clang (or clang-cl) due to the use of the Swift calling convention, we can use the `-Isystem` flag to indicate that the headers are system headers and should not generate warnings.  This is important as there are uses of anonymous unions and structs in the Windows headers which will trigger warnings which are treated as errors and thus break the build.